+++ Apps/Examples/Examples/All Examples/MultipleGeometriesExample.swift
@@ -36,7 +36,7 @@ public class MultipleGeometriesExample: UIViewController, ExampleProtocol {
 
         do {
             let data = try Data(contentsOf: filePath)
-            featureCollection = try GeoJSON.parse(FeatureCollection.self, from: data)
+            featureCollection = try JSONDecoder().decode(FeatureCollection.self, from: data)
         } catch {
             print("Error parsing data: \(error)")
         }