+++ Tests/MapboxMapsTests/Foundation/GeoJSON/FeatureCollectionTests.swift
@@ -3,17 +3,12 @@ import CoreLocation
 @testable import MapboxMaps
 
 // Disabling rules against force casting for test file.
-// swiftlint:disable explicit_top_level_acl explicit_acl force_try force_cast
+// swiftlint:disable explicit_top_level_acl explicit_acl force_try
 class FeatureCollectionTests: XCTestCase {
 
     func testFeatureCollection() {
         let data = try! Fixture.geojsonData(from: "featurecollection")!
-        let geojson = try! GeoJSON.parse(FeatureCollection.self, from: data)
-
-        XCTAssert(geojson.features[0].geometry.type == .LineString)
-        XCTAssert(geojson.features[1].geometry.type == .Polygon)
-        XCTAssert(geojson.features[2].geometry.type == .Polygon)
-        XCTAssert(geojson.features[3].geometry.type == .Point)
+        let geojson = try! JSONDecoder().decode(FeatureCollection.self, from: data)
 
         let lineStringFeature = geojson.features[0]
         guard case let .lineString(lineStringCoordinates) = lineStringFeature.geometry else {
@@ -21,7 +16,7 @@ class FeatureCollectionTests: XCTestCase {
             return
         }
         XCTAssert(lineStringCoordinates.coordinates.count == 19)
-        XCTAssert(lineStringFeature.properties!["id"] as! Int == 1)
+        XCTAssertEqual(lineStringFeature.properties?["id"], 1)
         XCTAssert(lineStringCoordinates.coordinates.first!.latitude == -26.17500493262446)
         XCTAssert(lineStringCoordinates.coordinates.first!.longitude == 27.977542877197266)
 
@@ -30,32 +25,31 @@ class FeatureCollectionTests: XCTestCase {
             XCTFail("Failed to create a Polygon.")
             return
         }
-        XCTAssert(polygonFeature.properties!["id"] as! Int == 2)
+        XCTAssertEqual(polygonFeature.properties?["id"], 2)
         XCTAssert(polygonCoordinates.coordinates[0].count == 21)
         XCTAssert(polygonCoordinates.coordinates[0].first!.latitude == -26.199035448897074)
         XCTAssert(polygonCoordinates.coordinates[0].first!.longitude == 27.972049713134762)
+
+        if case .polygon = geojson.features[2].geometry {} else {
+            XCTFail("Failed to create a Polygon.")
+        }
 
         let pointFeature = geojson.features[3]
         guard case let .point(pointCoordinates) = pointFeature.geometry else {
             XCTFail("Failed to create a Point.")
             return
         }
-        XCTAssert(pointFeature.properties!["id"] as! Int == 4)
+        XCTAssertEqual(pointFeature.properties?["id"], 4)
         XCTAssert(pointCoordinates.coordinates.latitude == -26.152510345365126)
         XCTAssert(pointCoordinates.coordinates.longitude == 27.95642852783203)
     }
 
     func testDecodedFeatureCollection() {
         let data = try! Fixture.geojsonData(from: "featurecollection")!
-        let geojson = try! GeoJSON.parse(FeatureCollection.self, from: data)
+        let geojson = try! JSONDecoder().decode(FeatureCollection.self, from: data)
 
         let encodedData = try! JSONEncoder().encode(geojson)
-        let decoded = try! GeoJSON.parse(FeatureCollection.self, from: encodedData)
-
-        XCTAssert(decoded.features[0].geometry.type == .LineString)
-        XCTAssert(decoded.features[1].geometry.type == .Polygon)
-        XCTAssert(decoded.features[2].geometry.type == .Polygon)
-        XCTAssert(decoded.features[3].geometry.type == .Point)
+        let decoded = try! JSONDecoder().decode(FeatureCollection.self, from: encodedData)
 
         let decodedLineStringFeature = decoded.features[0]
         guard case let .lineString(decodedLineStringCoordinates) = decodedLineStringFeature.geometry else {
@@ -63,7 +57,7 @@ class FeatureCollectionTests: XCTestCase {
                    return
                }
         XCTAssert(decodedLineStringCoordinates.coordinates.count == 19)
-        XCTAssert(decodedLineStringFeature.properties!["id"] as! Int == 1)
+        XCTAssertEqual(decodedLineStringFeature.properties?["id"], 1)
         XCTAssert(decodedLineStringCoordinates.coordinates.first!.latitude == -26.17500493262446)
         XCTAssert(decodedLineStringCoordinates.coordinates.first!.longitude == 27.977542877197266)
 
@@ -72,31 +66,39 @@ class FeatureCollectionTests: XCTestCase {
             XCTFail("Failed to decode Polygon.")
             return
         }
-        XCTAssert(decodedPolygonFeature.properties!["id"] as! Int == 2)
+        XCTAssertEqual(decodedPolygonFeature.properties?["id"], 2)
         XCTAssert(decodedPolygonCoordinates.coordinates[0].count == 21)
         XCTAssert(decodedPolygonCoordinates.coordinates[0].first!.latitude == -26.199035448897074)
         XCTAssert(decodedPolygonCoordinates.coordinates[0].first!.longitude == 27.972049713134762)
+
+        if case .polygon = decoded.features[2].geometry {} else {
+            XCTFail("Failed to decode a Polygon.")
+        }
 
         let decodedPointFeature = decoded.features[3]
         guard case let .point(decodedPointCoordinates) = decodedPointFeature.geometry else {
             XCTFail("Failed to decode Point.")
             return
         }
-        XCTAssert(decodedPointFeature.properties!["id"] as! Int == 4)
+        XCTAssertEqual(decodedPointFeature.properties?["id"], 4)
         XCTAssert(decodedPointCoordinates.coordinates.latitude == -26.152510345365126)
         XCTAssert(decodedPointCoordinates.coordinates.longitude == 27.95642852783203)
     }
 
     func testFeatureCollectionDecodeWithoutProperties() {
         let data = try! Fixture.geojsonData(from: "featurecollection-no-properties")!
-        let geojson = try! GeoJSON.parse(data)
-        XCTAssert(geojson.decoded is FeatureCollection)
+        let geojson = try! JSONDecoder().decode(GeoJSONObject.self, from: data)
+        if case .featureCollection = geojson {} else {
+            XCTFail("GeoJSON should decode as a feature collection.")
+        }
     }
 
     func testUnkownFeatureCollection() {
         let data = try! Fixture.geojsonData(from: "featurecollection")!
-        let geojson = try! GeoJSON.parse(data)
-        XCTAssert(geojson.decoded is FeatureCollection)
+        let geojson = try! JSONDecoder().decode(GeoJSONObject.self, from: data)
+        if case .featureCollection = geojson {} else {
+            XCTFail("GeoJSON should decode as a feature collection.")
+        }
     }
 
     func testPerformanceDecodeFeatureCollection() {
@@ -104,14 +106,14 @@ class FeatureCollectionTests: XCTestCase {
 
         measure {
             for _ in 0...100 {
-                _ = try! GeoJSON.parse(FeatureCollection.self, from: data)
+                _ = try! JSONDecoder().decode(FeatureCollection.self, from: data)
             }
         }
     }
 
     func testPerformanceEncodeFeatureCollection() {
         let data = try! Fixture.geojsonData(from: "featurecollection")!
-        let decoded = try! GeoJSON.parse(FeatureCollection.self, from: data)
+        let decoded = try! JSONDecoder().decode(FeatureCollection.self, from: data)
 
         measure {
             for _ in 0...100 {
@@ -125,7 +127,7 @@ class FeatureCollectionTests: XCTestCase {
 
         measure {
             for _ in 0...100 {
-                let decoded = try! GeoJSON.parse(FeatureCollection.self, from: data)
+                let decoded = try! JSONDecoder().decode(FeatureCollection.self, from: data)
                 _ = try! JSONEncoder().encode(decoded)
             }
         }