+++ Tests/MapboxMapsTests/Foundation/GeoJSON/LineStringTests.swift
@@ -2,16 +2,15 @@ import XCTest
 import CoreLocation
 @testable import MapboxMaps
 
-// swiftlint:disable explicit_top_level_acl explicit_acl force_try force_cast
+// swiftlint:disable explicit_top_level_acl explicit_acl force_try
 let metersPerMile: CLLocationDistance = 1_609.344
 
 class LineStringTests: XCTestCase {
 
     func testLineStringFeature() {
         let data = try! Fixture.geojsonData(from: "simple-line")!
-        let geojson = try! GeoJSON.parse(Feature.self, from: data)
+        let geojson = try! JSONDecoder().decode(Feature.self, from: data)
 
-        XCTAssert(geojson.geometry.type == .LineString)
         guard case let .lineString(lineStringCoordinates) = geojson.geometry else {
             XCTFail("Failed to create a LineString.")
             return
@@ -22,17 +21,17 @@ class LineStringTests: XCTestCase {
         let last = CLLocationCoordinate2D(latitude: 10, longitude: 0)
         XCTAssert(lineStringCoordinates.coordinates.first == first)
         XCTAssert(lineStringCoordinates.coordinates.last == last)
-        XCTAssert(geojson.identifier!.value as! String == "1")
+        XCTAssertEqual(geojson.identifier, "1")
 
         let encodedData = try! JSONEncoder().encode(geojson)
-        let decoded = try! GeoJSON.parse(Feature.self, from: encodedData)
+        let decoded = try! JSONDecoder().decode(Feature.self, from: encodedData)
         guard case let .lineString(decodedLineStringCoordinates) = decoded.geometry else {
             XCTFail("Failed to create a LineString.")
             return
         }
 
         XCTAssertEqual(lineStringCoordinates, decodedLineStringCoordinates)
-        XCTAssertEqual(geojson.identifier!.value as! String, decoded.identifier!.value! as! String)
+        XCTAssertEqual(geojson.identifier, decoded.identifier)
     }
 }
 //