+++ Tests/MapboxMapsTests/Foundation/GeoJSON/PolygonTests.swift
@@ -3,17 +3,17 @@ import CoreLocation
 @testable import MapboxMaps
 
 // Disabling rules against force casting for test file.
-// swiftlint:disable explicit_top_level_acl explicit_acl force_try force_cast
+// swiftlint:disable explicit_top_level_acl explicit_acl force_try
 class PolygonTests: XCTestCase {
 
     func testPolygonFeature() {
         let data = try! Fixture.geojsonData(from: "polygon")!
-        let geojson = try! GeoJSON.parse(Feature.self, from: data)
+        let geojson = try! JSONDecoder().decode(Feature.self, from: data)
 
         let firstCoordinate = CLLocationCoordinate2D(latitude: 37.00255267215955, longitude: -109.05029296875)
         let lastCoordinate = CLLocationCoordinate2D(latitude: 40.6306300839918, longitude: -108.56689453125)
 
-        XCTAssert((geojson.identifier!.value as! Number).value! as! Double == 1.01)
+        XCTAssertEqual(geojson.identifier, 1.01)
 
         guard case let .polygon(polygon) = geojson.geometry else {
             XCTFail("Failed to create polygon.")
@@ -25,14 +25,14 @@ class PolygonTests: XCTestCase {
         XCTAssert(polygon.innerRings.first?.coordinates.count == 5)
 
         let encodedData = try! JSONEncoder().encode(geojson)
-        let decoded = try! GeoJSON.parse(Feature.self, from: encodedData)
+        let decoded = try! JSONDecoder().decode(Feature.self, from: encodedData)
         guard case let .polygon(decodedPolygon) = decoded.geometry else {
-                   XCTFail("Failed to create polygon")
-                   return
-               }
+            XCTFail("Failed to create polygon")
+            return
+        }
 
         XCTAssertEqual(polygon, decodedPolygon)
-        XCTAssertEqual(geojson.identifier!.value as! Number, decoded.identifier!.value! as! Number)
+        XCTAssertEqual(geojson.identifier, decoded.identifier)
         XCTAssert(decodedPolygon.outerRing.coordinates.first == firstCoordinate)
         XCTAssert(decodedPolygon.innerRings.last?.coordinates.last == lastCoordinate)
         XCTAssert(decodedPolygon.outerRing.coordinates.count == 5)